feat: Game clock

+++ Assets/Framework/Runtime/Logic/IClock.cs
@@ -4,6 +4,6 @@
     {
         ulong FrameId { get; }
         ulong Time { get; }
-        ulong DeltaTime { get; }
+        uint DeltaTime { get; }
     }
 }
